Refactor RMQueryContext a bit
